complete udp

+++ server/server_main.cpp
@@ -24,7 +24,7 @@ int main(int argc, char **argv) {
         else if (command == "list") std::cout << server.list_clients() << std::endl;
         else if (command == "killall") server.close_all_clients();
         else if (!command.compare(0, 4, "kill")) {
-            auto&& client_id = std::stoi(command.substr(5));
+            auto&& client_id = std::stoll(command.substr(5));
             server.close_client(client_id);
         } else if (command == "shutdown") {
             break;
